Added Masks (now, it's better)

+++ Computer Vision/src/model/Normalizer.java
@@ -2,7 +2,7 @@ package model;
 
 public class Normalizer {
 	
-	public static int[][] normalizeImage(int[][] matrix){
+	public static int[][] normalizeImage(double[][] matrix){
 		
 		double min = matrix[0][0];
 		double max = matrix[0][0];
@@ -18,7 +18,7 @@ public class Normalizer {
 
 		for (int k = 0; k < matrix.length; k++) {
 			for (int k2 = 0; k2 < matrix[0].length; k2++) {
-				int grayPixel = matrix[k][k2];
+				double grayPixel = matrix[k][k2];
 				int newGrayPixel = 0;
 				
 				newGrayPixel = (int) (255 * (grayPixel - min) / (max - min));
